backend/requirements.txt
Flask
gunicorn
Flask-Cors
pandas
openpyxl
Flask-SocketIO
python-nmap
requests 